norm and corrections
