+++ out
@@ -1,15 +1,24 @@
-f1
-f2
-includes
-libft
-Makefile
-minishell
-minishell.excalidraw
-minishell_tester-main
-minishell_tester-main.zip
-mini_tests
-out
-readline.supp
-README.md
-src
-tests
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   minishell.h                                        :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: ighannam <ighannam@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/11/04 12:57:32 by brunofer          #+#    #+#             */
+/*   Updated: 2025/12/09 13:35:08 by ighannam         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#ifndef MINISHELL_H
+# define MINISHELL_H
+
+# include "../src/core/executer/executer.h"
+# include "../src/core/lexer/lexer.h"
+# include "../src/core/parser/parser.h"
+# include "../src/core/reader/reader.h"
+# include "../src/core/signals/signals.h"
+# include "../src/data_structures/data_structures.h"
+# include "../src/utils/utils.h"
+
+#endif
