+++ src/core/executer/and.c
@@ -6,7 +6,7 @@
 /*   By: ighannam <ighannam@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/11/21 13:12:21 by ighannam          #+#    #+#             */
-/*   Updated: 2025/12/08 20:24:41 by ighannam         ###   ########.fr       */
+/*   Updated: 2025/12/09 14:39:41 by ighannam         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
