+++ src/core/executer/builtins/ft_cd.c
@@ -6,7 +6,7 @@
 /*   By: ighannam <ighannam@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/11/26 11:31:58 by ighannam          #+#    #+#             */
-/*   Updated: 2025/12/08 18:58:55 by ighannam         ###   ########.fr       */
+/*   Updated: 2025/12/09 16:29:16 by ighannam         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,11 +24,15 @@ int	ft_cd(t_binary_tree_node *node)
 	ht = ft_find_ht(ft_get_ht_env(node), "PWD")->content;
 	old_pwd = ft_strdup(((t_env_value *)ht->value)->value);
 	if (ft_verify_cd(node) == 1)
+	{
+		free(old_pwd);
 		return (1);
+	}
 	pwd = getcwd(NULL, 0);
 	if (!pwd)
 	{
 		perror("getcwd");
+		free(old_pwd);
 		return (1);
 	}
 	ft_update_ht(node, pwd, old_pwd);
