+++ src/core/executer/here_doc_ii.c
@@ -6,7 +6,7 @@
 /*   By: ighannam <ighannam@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/11/28 16:39:37 by ighannam          #+#    #+#             */
-/*   Updated: 2025/12/08 20:29:23 by ighannam         ###   ########.fr       */
+/*   Updated: 2025/12/09 12:56:57 by ighannam         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
