+++ src/core/executer/mapper/ast_node_props.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ast_node_props.c                                   :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: valero <valero@student.42.fr>              +#+  +:+       +#+        */
+/*   By: ighannam <ighannam@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/11/28 17:54:45 by ighannam          #+#    #+#             */
-/*   Updated: 2025/12/08 15:59:26 by valero           ###   ########.fr       */
+/*   Updated: 2025/12/09 14:38:47 by ighannam         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,4 +26,14 @@ t_ast_node_type	ft_get_type(t_binary_tree_node *node)
 
 	ast_node = (t_ast_node *)(node->content);
 	return (ast_node->type);
+}
+
+void	free_ast_node(void *arg)
+{
+	t_ast_node	*node;
+
+	if (!arg)
+		return ;
+	node = (t_ast_node *)arg;
+	node->destroy(&node, ft_free_exec);
 }
