+++ src/core/signals/signals.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   signals.h                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: brunofer <brunofer@student.42.fr>          +#+  +:+       +#+        */
+/*   By: ighannam <ighannam@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/11/04 13:15:38 by brunofer          #+#    #+#             */
-/*   Updated: 2025/12/09 12:48:47 by brunofer         ###   ########.fr       */
+/*   Updated: 2025/12/09 13:39:14 by ighannam         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,9 +14,11 @@
 # define SIGNALS_H
 
 # include <signal.h>
+# include <stdio.h>
 # include <sys/ioctl.h>
+# include <unistd.h>
+# include <readline/readline.h>
 
-// void	ft_handle_signals(void);
 int		ft_get_sig(void);
 void	ft_set_sig(int value);
 void	ft_init_sig_parent(void);
@@ -26,6 +28,5 @@ void	ft_init_sig_ignore(void);
 void	ft_handle_sig_ignore(int sig);
 void	ft_init_sig_heredoc(void);
 void	ft_handle_sig_heredoc(int sig);
-// int		signs(void);
 
 #endif
